feat/Moving jobs list to 'Houses (live)' screen
