+++ src/containers/HouseWatchFilterContainer.js
@@ -1,10 +1,6 @@
 import React from 'react';
 import {
   ScrollView,
-  View,
-  Text,
-  StyleSheet,
-  Alert,
 } from 'react-native';
 import { withNavigation, NavigationEvents } from 'react-navigation';
 import PropTypes from 'prop-types';
@@ -13,26 +9,6 @@ import { connect } from 'react-redux';
 import HouseFilter from '../components/house/HouseFilter';
 import actions from '../store/actions';
 import * as Colors from '../constants/Colors';
-import HouseJob from '../components/house/HouseJob';
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: Colors.white,
-    justifyContent: 'center',
-    paddingHorizontal: 10,
-    flex: 1,
-  },
-  job: {
-    fontSize: 18,
-    paddingTop: 10,
-    paddingBottom: 10,
-    paddingLeft: 5,
-  },
-  divider: {
-    borderBottomColor: Colors.lightGray,
-    borderBottomWidth: 1,
-  },
-});
 
 class HouseWatchFilterContainer extends React.PureComponent {
   constructor(props) {
@@ -70,34 +46,6 @@ class HouseWatchFilterContainer extends React.PureComponent {
     changePriceTo(value);
   }
 
-  onCloseJob(value) {
-    const { removeJob } = this.props;
-    Alert.alert(
-      'Remove task',
-      'Are you sure about that?',
-      [
-        {
-          text: 'Cancel',
-          style: 'cancel',
-        },
-        {
-          text: 'OK',
-          onPress: () => removeJob(value),
-        },
-      ],
-      { cancelable: false },
-    );
-  }
-
-  onPlayPauseJob(value) {
-    const { pauseJob, resumeJob } = this.props;
-    if (value.state && value.state === 'RUNNING') {
-      pauseJob(value.id);
-    } else {
-      resumeJob(value.id);
-    }
-  }
-
   render() {
     const handlers = {
       roomsToHandler: this.onChangeHouseWatchFilterRoomsTo,
@@ -106,47 +54,15 @@ class HouseWatchFilterContainer extends React.PureComponent {
       priceFromHandler: this.onChangeHouseWatchFilterPriceFrom,
       applyFilter: this.onApplyHouseWatchFilter,
     };
-    const { navigation, filters, jobs } = this.props;
-    if (jobs.length !== 0) {
-      return (
-        <ScrollView style={{ backgroundColor: Colors.white }}>
-          <Text style={{ fontSize: 20, padding: 10, backgroundColor: Colors.silver }}>Active tasks:</Text>
-          <View style={styles.container}>
-            {jobs.map((value, index) => (
-              <HouseJob
-                key={`job-${index + 1}`}
-                text={`${index + 1}. ${value.jobId} [${value.params.rooms}] - $[${value.params.min}-${value.params.max}]`}
-                iosIcon={value.state === 'RUNNING' ? 'ios-pause' : 'ios-play'}
-                otherIcon={value.state === 'RUNNING' ? 'md-pause' : 'md-play'}
-                onPlayPause={() => this.onPlayPauseJob({ id: value.jobId, state: value.state })}
-                onClose={() => this.onCloseJob(value.jobId)}
-              />
-            ))}
-          </View>
-          <Text style={{ fontSize: 20, padding: 10, backgroundColor: Colors.silver }}>Create new task:</Text>
-          <HouseFilter filters={filters} handlers={handlers} />
-          <NavigationEvents
-            onDidBlur={() => navigation.pop(null)}
-          />
-        </ScrollView>
-      );
-    } else {
-      return (
-        <ScrollView style={{ backgroundColor: Colors.white }}>
-          <Text style={{ fontSize: 20, padding: 10, backgroundColor: Colors.silver }}>Started tasks:</Text>
-          <View style={styles.container}>
-            <Text style={styles.job}>
-              no active tasks
-            </Text>
-          </View>
-          <Text style={{ fontSize: 20, padding: 10, backgroundColor: Colors.silver }}>Create new task:</Text>
-          <HouseFilter filters={filters} handlers={handlers} />
-          <NavigationEvents
-            onDidBlur={() => navigation.pop(null)}
-          />
-        </ScrollView>
-      );
-    }
+    const { navigation, filters } = this.props;
+    return (
+      <ScrollView style={{ backgroundColor: Colors.white }}>
+        <HouseFilter filters={filters} handlers={handlers} />
+        <NavigationEvents
+          onDidBlur={() => navigation.pop(null)}
+        />
+      </ScrollView>
+    );
   }
 }
 
@@ -158,7 +74,6 @@ function mapStateToProps(state) {
       priceTo: state.houseFilterLiveReducers.priceTo,
       priceFrom: state.houseFilterLiveReducers.priceFrom,
     },
-    jobs: state.houseWatchLotsReducers.houseWatchJobs,
   };
 }
 
@@ -169,24 +84,17 @@ function mapDispatchToProps(dispatch) {
     changePriceFrom: value => dispatch(actions.houseWatchLotsFilterActions.updateHouseWatchFilterPriceFrom(value)),
     changePriceTo: value => dispatch(actions.houseWatchLotsFilterActions.updateHouseWatchFilterPriceTo(value)),
     applyFilter: value => dispatch(actions.houseWatchLotsFilterActions.updateHouseWatchFilterApply(value)),
-    removeJob: value => dispatch(actions.houseWatchLotsActions.removeHouseWatchJob(value)),
-    pauseJob: value => dispatch(actions.houseWatchLotsActions.pauseHouseWatchJob(value)),
-    resumeJob: value => dispatch(actions.houseWatchLotsActions.resumeHouseWatchJob(value)),
   };
 }
 
 HouseWatchFilterContainer.propTypes = {
   navigation: PropTypes.objectOf(PropTypes.any).isRequired,
   filters: PropTypes.objectOf(PropTypes.any).isRequired,
-  jobs: PropTypes.arrayOf(PropTypes.any).isRequired,
   changeRoomsTo: PropTypes.func.isRequired,
   changeRoomsFrom: PropTypes.func.isRequired,
   changePriceTo: PropTypes.func.isRequired,
   changePriceFrom: PropTypes.func.isRequired,
   applyFilter: PropTypes.func.isRequired,
-  removeJob: PropTypes.func.isRequired,
-  pauseJob: PropTypes.func.isRequired,
-  resumeJob: PropTypes.func.isRequired,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(withNavigation(HouseWatchFilterContainer));
