+++ src/containers/HouseWatchLotsContainer.js
@@ -1,23 +1,74 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import {
+  Alert,
   FlatList,
+  ScrollView,
   ActivityIndicator,
   View,
   Text,
   Switch,
+  StyleSheet,
 } from 'react-native';
 import PropTypes from 'prop-types';
 import actions from '../store/actions';
 
 import HouseLotCard from '../components/house/HouseLotCard';
 import BgMessage from '../components/bgmessage/BackgroundMessage';
+import HouseJob from '../components/house/HouseJob';
 import * as Colors from '../constants/Colors';
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: Colors.white,
+    justifyContent: 'center',
+    paddingHorizontal: 10,
+    flex: 1,
+  },
+  job: {
+    fontSize: 18,
+    paddingTop: 10,
+    paddingBottom: 10,
+    paddingLeft: 5,
+  },
+  divider: {
+    borderBottomColor: Colors.lightGray,
+    borderBottomWidth: 1,
+  },
+});
 
 class HouseWatchLotsContainer extends React.Component {
   componentDidMount() {
     const { onCheckHouseWatchState } = this.props;
     onCheckHouseWatchState();
+  }
+
+  onCloseJob(value) {
+    const { removeJob } = this.props;
+    Alert.alert(
+      'Remove task',
+      'Are you sure about that?',
+      [
+        {
+          text: 'Cancel',
+          style: 'cancel',
+        },
+        {
+          text: 'OK',
+          onPress: () => removeJob(value),
+        },
+      ],
+      { cancelable: false },
+    );
+  }
+
+  onPlayPauseJob(value) {
+    const { pauseJob, resumeJob } = this.props;
+    if (value.state && value.state === 'RUNNING') {
+      pauseJob(value.id);
+    } else {
+      resumeJob(value.id);
+    }
   }
 
   render() {
@@ -26,10 +77,12 @@ class HouseWatchLotsContainer extends React.Component {
       houseWatchLots,
       isWatching,
       onUpdateHouseWatchState,
+      jobs,
+      isEditing,
     } = this.props;
     if (!houseWatchLots.length && isFetching) return <ActivityIndicator size="large" color={Colors.lightGray} />;
     return (
-      <View>
+      <View style={{ flexDirection: 'column' }}>
         <View style={{ display: 'flex', flexDirection: 'row', padding: 10 }}>
           <Text style={{ fontSize: 24, color: Colors.gray, marginLeft: 9 }}>Live tracking</Text>
           <Switch
@@ -39,7 +92,22 @@ class HouseWatchLotsContainer extends React.Component {
             onValueChange={onUpdateHouseWatchState}
           />
         </View>
-        <FlatList
+        <ScrollView style={{ backgroundColor: Colors.white, marginBottom: 50 }}>
+          <View style={styles.container}>
+            {jobs.map((value, index) => (
+              <HouseJob
+                key={`job-${index + 1}`}
+                text={`${index + 1}. ${value.jobId} [${value.params.rooms}] - $[${value.params.min}-${value.params.max}]`}
+                iosIcon={value.state === 'RUNNING' ? 'ios-pause' : 'ios-play'}
+                otherIcon={value.state === 'RUNNING' ? 'md-pause' : 'md-play'}
+                onPlayPause={() => this.onPlayPauseJob({ id: value.jobId, state: value.state })}
+                onClose={() => this.onCloseJob(value.jobId)}
+                isEditing={isEditing}
+              />
+            ))}
+          </View>
+        </ScrollView>
+        {/* <FlatList
           style={{ marginBottom: 30 }}
           data={houseWatchLots}
           renderItem={({ item }) => <HouseLotCard item={item} />}
@@ -48,7 +116,7 @@ class HouseWatchLotsContainer extends React.Component {
           onEndReachedThreshold={0}
           refreshing={isFetching}
           ListEmptyComponent={() => <BgMessage text="There are no new houses" />}
-        />
+        /> */}
       </View>
     );
   }
@@ -58,10 +126,12 @@ function mapStateToProps(state) {
   return {
     isFetching: state.houseWatchLotsReducers.isFetching,
     isWatching: state.houseWatchLotsReducers.isWatching,
+    isEditing: state.houseWatchLotsReducers.isEditing,
     houseWatchLots: state.houseWatchLotsReducers.houseWatchLots,
     page: state.houseWatchLotsReducers.page,
     itemsPerPage: state.houseWatchLotsReducers.itemsPerPage,
     error: state.houseWatchLotsReducers.error ? state.houseWatchLotsReducers.error : null,
+    jobs: state.houseWatchLotsReducers.houseWatchJobs,
   };
 }
 
@@ -69,15 +139,23 @@ function mapDispatchToProps(dispatch) {
   return {
     onUpdateHouseWatchState: value => dispatch(actions.houseWatchLotsActions.watchHouseLots(value)),
     onCheckHouseWatchState: value => dispatch(actions.houseWatchLotsActions.checkWatchHouseLotsState(value)),
+    removeJob: value => dispatch(actions.houseWatchLotsActions.removeHouseWatchJob(value)),
+    pauseJob: value => dispatch(actions.houseWatchLotsActions.pauseHouseWatchJob(value)),
+    resumeJob: value => dispatch(actions.houseWatchLotsActions.resumeHouseWatchJob(value)),
   };
 }
 
 HouseWatchLotsContainer.propTypes = {
   isFetching: PropTypes.bool.isRequired,
   isWatching: PropTypes.bool.isRequired,
+  isEditing: PropTypes.bool.isRequired,
   houseWatchLots: PropTypes.arrayOf(PropTypes.any).isRequired,
+  jobs: PropTypes.arrayOf(PropTypes.any).isRequired,
   onUpdateHouseWatchState: PropTypes.func.isRequired,
   onCheckHouseWatchState: PropTypes.func.isRequired,
+  removeJob: PropTypes.func.isRequired,
+  pauseJob: PropTypes.func.isRequired,
+  resumeJob: PropTypes.func.isRequired,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(HouseWatchLotsContainer);
